File system revision, schools list updated and styled, logos added, fonts manrope and manrope-bold added
