+++ src/assets/imgs/skillsIcons/javascript.png
