+++ src/assets/imgs/sliderphotos/photoImg2.jpg
