+++ src/data/skills.js
@@ -1,0 +1,89 @@
+import adobe from "../assets/imgs/skillsIcons/adobe.png";
+import bootstrap from "../assets/imgs/skillsIcons/bootstrap.png";
+import css from "../assets/imgs/skillsIcons/css.png";
+import express from "../assets/imgs/skillsIcons/express.png";
+import git from "../assets/imgs/skillsIcons/git.png";
+import github from "../assets/imgs/skillsIcons/github.png";
+import heroku from "../assets/imgs/skillsIcons/adobe.png";
+import html from "../assets/imgs/skillsIcons/html.png";
+import javascript from "../assets/imgs/skillsIcons/javascript.png";
+import knex from "../assets/imgs/skillsIcons/knex.png";
+import mochaChai from "../assets/imgs/skillsIcons/mochaChai.png";
+import npm from "../assets/imgs/skillsIcons/npm.png";
+import postgresql from "../assets/imgs/skillsIcons/postgresql.png";
+import react from "../assets/imgs/skillsIcons/react.png";
+import restful from "../assets/imgs/skillsIcons/restful.png";
+import vercel from "../assets/imgs/skillsIcons/vercel.png";
+
+const skills = [
+  { 
+    name: "Adobe Creative Suite",
+    img: adobe,
+  },
+  { 
+    name: "Bootstrap.js",
+    img: bootstrap,
+  },
+  { 
+    name: "Css",
+    img: css,
+    abbr: "Cascading Style Sheets",
+  },
+  { 
+    name: "Express.js",
+    img: express,
+  },
+  { 
+    name: "Git",
+    img: git,
+  },
+  { 
+    name: "GitHub",
+    img: github,
+  },
+  { 
+    name: "Heroku",
+    img: heroku,
+  },
+  { 
+    name: "Html",
+    img: html,
+    abbr: "HyperText Markup Language"
+  },
+  { 
+    name: "Javascript",
+    img: javascript,
+  },
+  { 
+    name: "Knex.js",
+    img: knex,
+  },
+  { 
+    name: "Mocha & Chai",
+    img: mochaChai,
+  },
+  { 
+    name: "NPM",
+    img: npm,
+    abbr: "Node Package Manager",
+  },
+  { 
+    name: "PostgreSQL",
+    img: postgresql,
+  },
+  { 
+    name: "React.js",
+    img: react,
+  },
+  { 
+    name: "RESTful API",
+    img: restful,
+    abbr: "Representational State Transfer"
+  },
+  { 
+    name: "Vercel",
+    img: vercel,
+  },
+];
+
+export default skills ;
