+++ src/data/sliderData.js
@@ -1,0 +1,30 @@
+import img1 from "../assets/imgs/sliderphotos/photoImg1.jpg";
+import img2 from "../assets/imgs/sliderphotos/photoImg2.jpg";
+import img3 from "../assets/imgs/sliderphotos/photoImg3.jpg";
+import img4 from "../assets/imgs/sliderphotos/photoImg4.jpg";
+import img5 from "../assets/imgs/sliderphotos/photoImg5.jpg";
+
+const sliderData = [
+  {
+    id: 1,
+    img: img1,
+  },
+  {
+    id: 2,
+    img: img2,
+  },
+  {
+    id: 3,
+    img: img3,
+  },
+  {
+    id: 4,
+    img: img4,
+  },
+  {
+    id: 5,
+    img: img5,
+  },
+];
+
+export default sliderData;
