+++ src/layout/aboutme/aboutme.css
@@ -13,6 +13,9 @@
 .mike__title {
   margin: 0;
 }
+.mike__about {
+  font-family: manrope;
+}
 
 .about-me__wrapper {
   margin: 0;
@@ -44,60 +47,6 @@
   z-index: 5;
 }
 
-.miniNav{
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  border-top: solid 2px var(--project__button-border-color);
-  border-bottom: solid 2px var(--project__button-border-color);
-}
-.miniNav__buttons-wrapper {
-  box-sizing:border-box;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-}
-.miniNav__button, .miniNav__button:link, .miniNav__button:visited {
-  margin: 0 5%;
-  background: none;
-  border: transparent;
-  color: var(--primary-font-color);
-  border-radius: 8px;
-  transition: 500ms;
-  text-transform: uppercase;
-  cursor:pointer;
-}
-.miniNav__button:focus:not(.focus-visible) {
-  outline: none;
-}
-.miniNav__button:hover, .miniNav__button:focus-visible {
-  outline-style: none;
-  box-shadow: none;
-  border-color: transparent;
-  color: var(--nav-active-font-color);
-  transition: 300ms;
-}
-.miniNav__button:active {
-  color: var(--mike-font);
-  transition: 0ms;
-}
-
-.info-wrapper {
-  position: relative;
-  overflow: visible;
-  min-height: 325px;
-  z-index: 90;
-  background-color: var(--lighter-bg-color);
-}
-.info-wrapper__padding {
-  padding: 2%;
-}
-.info-wrapper.block {
-  padding-left: 0;
-  padding-right: 0;
-}
-
 .active-screen{
   position: relative;
   top:0%;
@@ -111,48 +60,6 @@
   opacity: 0;
   display: none;
   transition: opacity ease-in-out 0.2s;
-}
-
-.school {
-  display: grid;
-  grid-template-columns: 20% auto;
-  margin: 3% 0;
-}
-.school__title {
-  color: var(--title-color);
-  font-size: 110%;
-  vertical-align: bottom;
-}
-.school__info {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-}
-.school a {
-  color: inherit;
-  transition: 500ms;
-  font-style: italic;
-}
-.school a:hover {
-  color: var(--nav-active-font-color);
-  transition: 300ms;
-}
-.school a:active{
-  color: var(--mike-font);
-  transition: 0ms;
-}
-.school a:focus:not(.focus-visible)  {
-  color: var(--nav-active-font-color);
-  transition: 300ms;
-  outline: none;
-}
-
-@media (min-width: 900px) {
-
-  .info-wrapper.block {
-    padding: 0;
-  }
-
 }
 
 @media (min-width: 1200px){
@@ -183,10 +90,6 @@
     display: none;
   }
 
-  .school {
-    grid-template-columns: 10% auto;
-  }
-
   .about-me__wrapper {
     margin: 0 auto;
     margin-bottom: 20px;
@@ -208,10 +111,6 @@
 
   .top-corner {
     border-radius: 0 25px 0 0;
-  }
-
-  .info-wrapper.block {
-    padding: 0;
   }
 
   .profile-image {
