+++ src/layout/aboutme/components/mininav/mininav.css
@@ -1,0 +1,38 @@
+.miniNav{
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  border-top: solid 2px var(--project__button-border-color);
+  border-bottom: solid 2px var(--project__button-border-color);
+}
+.miniNav__buttons-wrapper {
+  box-sizing:border-box;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+.miniNav__button, .miniNav__button:link, .miniNav__button:visited {
+  margin: 0 5%;
+  background: none;
+  border: transparent;
+  color: var(--primary-font-color);
+  border-radius: 8px;
+  transition: 500ms;
+  text-transform: uppercase;
+  cursor:pointer;
+}
+.miniNav__button:focus:not(.focus-visible) {
+  outline: none;
+}
+.miniNav__button:hover, .miniNav__button:focus-visible {
+  outline-style: none;
+  box-shadow: none;
+  border-color: transparent;
+  color: var(--nav-active-font-color);
+  transition: 300ms;
+}
+.miniNav__button:active {
+  color: var(--mike-font);
+  transition: 0ms;
+}
