--- src/layout/aboutme/components/photoSlider/data/sliderData.js
@@ -1,30 +1,0 @@
-import img1 from "../photos/photoImg1.jpg"
-import img2 from "../photos/photoImg2.jpg"
-import img3 from "../photos/photoImg3.jpg"
-import img4 from "../photos/photoImg4.jpg"
-import img5 from "../photos/photoImg5.jpg"
-
-const sliderData = [
-  {
-    id: 1,
-    img: img1,
-  },
-  {
-    id: 2,
-    img: img2,
-  },
-  {
-    id: 3,
-    img: img3,
-  },
-  {
-    id: 4,
-    img: img4,
-  },
-  {
-    id: 5,
-    img: img5,
-  },
-];
-
-export default sliderData;
