+++ src/layout/footer/footer.css
@@ -4,6 +4,7 @@
   text-align: end;
 }
 .footer__copyright {
+  font-family: manrope-bold;
   font-style: italic;
   font-size: 0.8rem;
   margin-right: 1%;
