+++ src/layout/navbar/navbar.css
@@ -8,6 +8,7 @@
   height: 7%;
   width: 100%;
   z-index: 9999;
+  font-family: manrope;
 }
 .navbar__buttons-wrapper {
   box-sizing:border-box;
