+++ src/layout/skills/Skills.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useInView } from "react-intersection-observer";
 import "./skills.css";
-import skills from "./data/skills";
+import skills from "../../data/skills";
 
 export default function Skills(){
   const [skillVisible, setSkillVisible] = useState(false);
@@ -21,9 +21,14 @@ export default function Skills(){
         <div className="skills-list" ref={skillRef}>
           {skills.map((skill, i) => {
             return (
-              <div className={`skill-card ${skillVisible ? `skill-load-in` : ""}`} key={i}>
+              <div className={`skill-card ${skillVisible ? "skill-load-in" : ""}`} key={i}>
                 <div><img className="skill-card__image" src={skill.img} alt={skill.name}/></div>
-                <div>{`${skill.name}`}</div>
+                {
+                  skill.abbr ? <div><abbr title={skill.abbr}>{`${skill.name}`}</abbr></div>
+                  : 
+                  <div>{`${skill.name}`}</div>
+                }
+                
               </div>
             );
           })}
