--- src/layout/skills/data/skills.js
@@ -1,85 +1,0 @@
-import adobe from "../imgs/adobe.png";
-import bootstrap from "../imgs/bootstrap.png";
-import css from "../imgs/css.png";
-import express from "../imgs/express.png";
-import git from "../imgs/git.png";
-import github from "../imgs/github.png";
-import heroku from "../imgs/adobe.png";
-import html from "../imgs/html.png";
-import javascript from "../imgs/javascript.png";
-import knex from "../imgs/knex.png";
-import mochaChai from "../imgs/mochaChai.png";
-import npm from "../imgs/npm.png";
-import postgresql from "../imgs/postgresql.png";
-import react from "../imgs/react.png";
-import restful from "../imgs/restful.png";
-import vercel from "../imgs/vercel.png";
-
-const skills = [
-  { 
-    name: "Adobe Creative Suite",
-    img: adobe,
-  },
-  { 
-    name: "Bootstrap.js",
-    img: bootstrap,
-  },
-  { 
-    name: "Css",
-    img: css,
-  },
-  { 
-    name: "Express.js",
-    img: express,
-  },
-  { 
-    name: "Git",
-    img: git,
-  },
-  { 
-    name: "GitHub",
-    img: github,
-  },
-  { 
-    name: "Heroku",
-    img: heroku,
-  },
-  { 
-    name: "Html",
-    img: html,
-  },
-  { 
-    name: "Javascript",
-    img: javascript,
-  },
-  { 
-    name: "Knex.js",
-    img: knex,
-  },
-  { 
-    name: "Mocha & Chai",
-    img: mochaChai,
-  },
-  { 
-    name: "NPM",
-    img: npm,
-  },
-  { 
-    name: "PostgreSQL",
-    img: postgresql,
-  },
-  { 
-    name: "React.js",
-    img: react,
-  },
-  { 
-    name: "RESTful API",
-    img: restful,
-  },
-  { 
-    name: "Vercel",
-    img: vercel,
-  },
-];
-
-export default skills ;
